Safari-611.1.10/LayoutTests/http/wpt/cache-storage/cache-quota-after-restart.any-expected.txt

CONSOLE MESSAGE: Cache API operation failed: Quota exceeded
CONSOLE MESSAGE: Cache API operation failed: Quota exceeded

PASS Increasing quota
PASS After network process restart, verify quota is computed according space being used
PASS After network process restart, verify quota is computed according space being used and does not increase
